backend/requirements.txt
flask>=2.3.0
flask-cors==4.0.0
python-dotenv>=1.0.0
pypdf>=3.15.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
rank-bm25>=0.2.2
chromadb>=0.4.0
openai>=1.0.0
werkzeug>=2.3.0
torch>=2.0.0
PyJWT>=2.10.0
Flask-Limiter>=4.0.0
#python-magic>=0.4.27         # for linux
python-magic-bin>=0.4.14    # for windows
pytest>=7.0.0
pytest-cov>=4.0.0
freezegun>=1.2.0
mcp>=0.9.0
ddgs>=2.0.0  # Renamed from duckduckgo-search
tavily-python>=0.5.0
langsmith>=0.1.0
psycopg2-binary>=2.9.0
psycopg[binary,pool]>=3.0.0  # Required for langgraph-checkpoint-postgres (pool for ConnectionPool)
redis>=5.0.0
prisma>=0.11.0
asgiref>=3.11.0
# LangGraph for agentic RAG implementation
langgraph>=0.2.0
langgraph-checkpoint-postgres>=0.0.1
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
ragas>=0.2.0
datasets>=2.0.0
pypdf2>=3.0.0
openpyxl>=3.0.0
reportlab>=4.4.0
markdown2>=2.4.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
dishka>=0.8.0
